script service and alert ulti and fix bug
