+++ QuanLyThuVien/src/main/java/com/bros/quanlythuvien/repository/impl/BookRepositoryImpl.java
@@ -13,12 +13,9 @@ import com.bros.quanlythuvien.utils.QueryBuilderUtils;
 import com.bros.quanlythuvien.utils.ValidateUtils;
 import java.sql.Connection;
 import java.sql.PreparedStatement;
-import java.sql.ResultSet;
 import java.sql.SQLException;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import javafx.scene.control.Alert;
 
 /**
  *
