+++ QuanLyThuVien/src/main/java/com/bros/quanlythuvien/repository/impl/LoanslipRepositoryImpl.java
@@ -23,6 +23,7 @@ import java.util.logging.Level;
 import java.util.logging.Logger;
 import javafx.scene.control.Alert;
 import com.bros.quanlythuvien.repository.LoanSlipRepository;
+import com.bros.quanlythuvien.utils.MessageBoxUtils;
 
 /**
  *
@@ -66,11 +67,7 @@ public class LoanSlipRepositoryImpl extends CommonRepositoryImpl<LoanSlipEntity>
 
                 long daysBetween = (sqlDate.getTime() - sqlTimestamp.getTime()) / (1000 * 60 * 60 * 24);
                 if (daysBetween > 0) {
-                    Alert alert = new Alert(Alert.AlertType.ERROR);
-                    alert.setTitle("INFORMATION");
-                    alert.setHeaderText("INFORMATION");
-                    alert.setContentText("Bạn đã trễ hạn " + daysBetween + " ngày và tiền phạt là: " + 5000 * daysBetween + " VNĐ");
-                    alert.showAndWait();
+                    MessageBoxUtils.AlertBox("ERROR", "Bạn đã trễ hạn " + daysBetween + " ngày và tiền phạt là: " + 5000 * daysBetween + " VNĐ", Alert.AlertType.ERROR);
                 }
             }
         } catch (SQLException e) {
@@ -274,17 +271,9 @@ public class LoanSlipRepositoryImpl extends CommonRepositoryImpl<LoanSlipEntity>
                         int rowsInserted = statement.executeUpdate();
                         if (rowsInserted > 0) {
                             deleteQuantity(book.getQuantity(), book.getId());
-                            Alert alert = new Alert(Alert.AlertType.INFORMATION);
-                            alert.setTitle("Informatin");
-                            alert.setHeaderText("Success");
-                            alert.setContentText("Thêm thành công");
-                            alert.showAndWait();
+                            MessageBoxUtils.AlertBox("INFORMATION", "Thêm thành công", Alert.AlertType.INFORMATION);
                         } else {
-                            Alert alert = new Alert(Alert.AlertType.ERROR);
-                            alert.setTitle("ERROR");
-                            alert.setHeaderText("ERROR");
-                            alert.setContentText("Thêm thất bại");
-                            alert.showAndWait();
+                            MessageBoxUtils.AlertBox("ERROR", "Thêm thất bạii", Alert.AlertType.ERROR);
                         }
                     } catch (Exception e) {
                         e.printStackTrace();
@@ -304,20 +293,12 @@ public class LoanSlipRepositoryImpl extends CommonRepositoryImpl<LoanSlipEntity>
                         }
                     }
                 } else {
-                    Alert alert = new Alert(Alert.AlertType.ERROR);
-                    alert.setTitle("Error");
-                    alert.setHeaderText("Error");
-                    alert.setContentText("Thư viện không đủ sách");
-                    alert.showAndWait();
+                    MessageBoxUtils.AlertBox("ERROR", "Thư viện không đủ sách", Alert.AlertType.ERROR);
                 }
 
             }
         } else {
-            Alert alert = new Alert(Alert.AlertType.ERROR);
-            alert.setTitle("Error");
-            alert.setHeaderText("Error");
-            alert.setContentText("Khách hàng không tồn tại hoặc bạn chưa thêm sách để tạo phiếu mượn");
-            alert.showAndWait();
+            MessageBoxUtils.AlertBox("ERROR", "Khách hàng không tồn tại hoặc bạn chưa thêm sách để tạo phiếu mượn", Alert.AlertType.ERROR);
         }
     }
 
@@ -464,9 +445,9 @@ public class LoanSlipRepositoryImpl extends CommonRepositoryImpl<LoanSlipEntity>
         }
 
     }
-    
+
     @Override
-    public  List<ReportModel> getReportBorrow(){
+    public List<ReportModel> getReportBorrow() {
         List<ReportModel> reports = new ArrayList<>();
         try {
             Connection conn = getConnection();
@@ -476,7 +457,7 @@ public class LoanSlipRepositoryImpl extends CommonRepositoryImpl<LoanSlipEntity>
             ResultSet rs = pstmt.executeQuery();
 
             // Xử lý dữ liệu lấy được ở đây
-            while(rs.next()) {
+            while (rs.next()) {
                 ReportModel report = new ReportModel();
                 report.setYear(rs.getInt("Year"));
                 report.setQuarter(rs.getString("Quarter"));
@@ -494,37 +475,36 @@ public class LoanSlipRepositoryImpl extends CommonRepositoryImpl<LoanSlipEntity>
         }
         return reports;
     }
-    
-    
+
     @Override
-    public  List<ReportModel> getReportReturn() {
+    public List<ReportModel> getReportReturn() {
         List<ReportModel> reports = new ArrayList<>();
-    try {
-        Connection conn = getConnection();
-        String sql = "SELECT YEAR(BorrowedDate) as Year, CONCAT('Qúy ', QUARTER(BorrowedDate)) AS Quarter, "
-                + "SUM(quantity) as NumReturnedBooks FROM loanslip WHERE isReturned = 1 "
-                + "GROUP BY YEAR(BorrowedDate), QUARTER(BorrowedDate)";
-        PreparedStatement pstmt = conn.prepareStatement(sql);
-        ResultSet rs = pstmt.executeQuery();
+        try {
+            Connection conn = getConnection();
+            String sql = "SELECT YEAR(BorrowedDate) as Year, CONCAT('Qúy ', QUARTER(BorrowedDate)) AS Quarter, "
+                    + "SUM(quantity) as NumReturnedBooks FROM loanslip WHERE isReturned = 1 "
+                    + "GROUP BY YEAR(BorrowedDate), QUARTER(BorrowedDate)";
+            PreparedStatement pstmt = conn.prepareStatement(sql);
+            ResultSet rs = pstmt.executeQuery();
 
-        // Xử lý dữ liệu lấy được ở đây
-        while(rs.next()) {
-            ReportModel report = new ReportModel();
-            report.setYear(rs.getInt("Year"));
-            report.setQuarter(rs.getString("Quarter"));
-            report.setQuantity(rs.getInt("NumReturnedBooks"));
-            // Thêm đối tượng report vào danh sách hoặc làm gì đó khác với dữ liệu này tại đây
-            reports.add(report);
+            // Xử lý dữ liệu lấy được ở đây
+            while (rs.next()) {
+                ReportModel report = new ReportModel();
+                report.setYear(rs.getInt("Year"));
+                report.setQuarter(rs.getString("Quarter"));
+                report.setQuantity(rs.getInt("NumReturnedBooks"));
+                // Thêm đối tượng report vào danh sách hoặc làm gì đó khác với dữ liệu này tại đây
+                reports.add(report);
+            }
+
+            // Đóng kết nối, PreparedStatement và ResultSet
+            rs.close();
+            pstmt.close();
+            conn.close();
+        } catch (Exception e) {
+            e.printStackTrace();
         }
-
-        // Đóng kết nối, PreparedStatement và ResultSet
-        rs.close();
-        pstmt.close();
-        conn.close();
-    } catch (Exception e) {
-        e.printStackTrace();
-    }
-    return reports;
+        return reports;
     }
 
 }
