+++ QuanLyThuVien/src/main/java/com/bros/quanlythuvien/service/CategoryService.java
@@ -6,17 +6,16 @@ package com.bros.quanlythuvien.service;
 
 import com.bros.quanlythuvien.model.CategoryModel;
 import java.util.List;
-import java.util.Map;
-import javafx.scene.control.ComboBox;
+
 
 /**
  *
  * @author phu nguyen
  */
 public interface CategoryService {
-        public List<CategoryModel> findAll();
-        
-        public CategoryModel findById(Integer id);
-         public void loadCate(ComboBox <String> availableBooks_category,Map<Integer, String> categoriesMap);
-        
+
+    public List<CategoryModel> findAll();
+
+    public CategoryModel findById(Integer id);
+
 }
